feat: separated getting vehicles from status request
